+++ docs/changelog.md
@@ -7,3 +7,23 @@ add script to support local .tmpl and .dspy file translation. it can build dynam
 * uses weakref to collect all the i18n widgets in i18n.py
 * show video play position
 
+## version 0.4.5
+* add Behavior handle when build Widget by blocks
+* support following behaviors
+1 ButtonBehavior
+2 CodeNavigationBehavior
+3 CompoundSelectionBehavior
+4 CoverBehavior
+5 DragBehavior
+6 EmacsBehavior
+7 FocusBehavior
+8 ToggleButtonBehavior
+9 TouchRippleBehavior
+10 TouchRippleButtonBehavior
+11 BGColorBehavior
+12 ModalBehavior
+13 swipeBehavior
+14 videoBehavior
+* DataGrid class add checkbox support
+
+